Implement pp6calculator Homework Exercise Part 2
Add operation to calculate x-intercept of straight line (y=mx+c).
Provide a new menu item which when selected asks the user for
the y-intercept and gradient of the line. Pass these numbers to
new intercept function that solves y=mx+c for the y=0 case. Output
the returned value of x to stdout.

If a zero (i.e. flat) gradient is entered, report a divide by zero
error and continue back to menu.