Implement pp6calculator Homework Exercise Part 3
Add an operation to solve for the roots of quadratic equations in the
case that the roots are real. Implement a quadratic function that takes
the coefficients and a boolean which selects whether to return the
positive or negative root.

Add a new item to the operations menu which when selected queries the
user for the quadratic coefficients. Check for two main error cases:

- Imaginary roots, in which warn that no solutions are possible (well,
in this implementation anyway).
- Zero quadratic coefficient.

Report errors in both cases before continuing to menu. Otherwise,
call quadratic function with coefficients, once for the positive
root and once for the negative root.

Add special result output case for the quadratic operation to output
both roots. Output other operation results as before.