Implement pp6calculator Homework Exercise Part 4
Add operations for calculating the lengths of 3- and 4-vectors from
their components. In each operation, query user for the vector
components in turn, passing these to the appropriate length calculating
function and outputting the result to stdout.

In the 4-vector case, report an error if the space-like part of the
vector is larger than the time-like part and continue to the menu.