Implement pp6calculator Homework Exercise Part 5
Provide operation for calculating invariant mass of two particles.
Query user for the components of each particle's 4-vector and pass
these to new inv_mass function if both vectors are timelike. If either
vector is spacelike, report and error and return to menu.

Implement inv_mass function in terms of the length() function for
4-Vectors to reuse code.