Implement pp6calculator Day 2 Exercise 1 solution
Implement new swap function for swapping values of two double variables.
Provide an operation in the menu to perform a swap, and add the call
to the swap function in the conditional block for operations taking
two numbers.

Report the values of the variables before and after the call two swap.
As this means swap does not have a result, exclude the operation from
having output reported at the end of the while block for the
operations.