Implement pp6calculator Day 2 Homework Part 2
Add the following support functions in PP6Math.hpp

- getRandom  : return a pseudo-random number between 0 and 1
- sumVectors : sum components of a vector supplied as three double
  arrays, returning the summed components
- getMeanAndStdDev : calculate the mean and standard deviation of an
  array of doubles
- swap : swap function for ints rather than doubles
- associative_sort : bubble sort of a an array of doubles, but returning
  an array of the indices of the sorted elements. The supplied array is
  not sorted.

Add operation to PP6Day2Menu to read in an integer N from the user
between 1 and 1000. Randomly generate N 4-vectors, first generating
random momentum components and a mass, from which the energy is derived.
Calculate and report the total momentum, mean and standard deviation of
the energy and use associative_sort to find the highest energy vector
and the index location in the original array.

Take care with error checking so that if any step in the analysis fails,
the dynamic arrays created are deleted cleanly.