Implement pp6calculator Day 2 Homework Part 3
Implement solution using a new function pp6day2_muonanalysis called as
an operation from the pp6day2_menu function.

Read in filename to be analysed from user, and use FileReader to check
file is valid and to count the number of mu+ and mu- particles observed
in "run4.dat". Create arrays of these sizes to hold the event number,
momentum components and energy of the muons and antimuons respectively.
Use a second FileReader instance to reiterate over the file and fill the
arrays with the relevant information. Hard code the value for the muon
mass to calculate mu-/mu+ energies from the extracted momenta.

Create arrays to hold invariant masses formed between each mu- with each
mu+, filling this and an indexing array using a bi-directional scheme
so that each entry in the index can go back to the indices of the
individual muon and antimuon indices.

Sort the invariant mass array using an associative sort to sort the
indexing array.

Read and present to stdout the 10 highest invariant masses and the event
number, energy and momenta of the mu+/mu- pair forming them.