Implement pp6calculator Day 3 Exercise 4 solution
Hide FourVector struct implementation using an Opaque Pointer, providing
`createFourVector` (including variant to construct with components) and
`destroyFourVector` functions for instance management.

Refactor `interval` and `boost_z` functions to use pointer-to-FourVector
and change menu items to use the new interface.

Provide a `userCreateFourVector` function to return a new instance of a
FourVector created by reading in the components from the user via cin.