Implement pp6calculator Day 3 Homework Part 1
Create new `ThreeVector` class to factor spatial vector functionality
out of `FourVector`. Provide same operator system as for `FourVector`.

Refactor `FourVector` to hold an instance of `ThreeVector` to represent
the space-like component. Reimplement methods and operators to use this
instance as needed.

Change value of `FourVector::c` to be unity to ensure natural units are
used for upcoming work with particle enrgy and momenta.