Implement pp6calculator Day 4 Exercise 4 solution
Implement ParticleInfo class for querying properties of a particle by
its PDG name or code. Provide public interface as per design discussion
on day 4. Use maps to provide lookup tables against PDG name or code.
Construction takes a .dbt file as input to fille the tables. If the
file is not readable or of the wrong format, the database will be empty.
Use the `size` method to indicate an empty, and hence invalid, database

Add operation to Day 4 menu calling new `pp6day4_check_particleinfo`
function. Inside function, query user for .dbt file and instantiate a
ParticleInfo instance using this file. Perform simple checks on the
resulting instance to test that it contains entries, and that
information on muons and antimuons is returned correctly.